feat(persistence): update product discount type and add entity configurations with seeding

+++ Core/EcommerceApi.Domain/Entities/Product.cs
@@ -13,7 +13,7 @@ namespace EcommerceApi.Domain.Entities
         public required string Description { get; set; }
         public required int BrandId { get; set; }
         public required decimal Price { get; set; }
-        public required int Discount { get; set; }
+        public required decimal Discount { get; set; }
         public Brand Brand { get; set; }
         public ICollection<Category> Categories { get; set; }
     }
